chore: removal of user job preference
