+++ __tests__/users.ts
@@ -146,12 +146,6 @@ import {
   DEFAULT_NOTIFICATION_SETTINGS,
   NotificationType,
 } from '../src/notifications/common';
-import { UserWorkExperience } from '../src/entity/user/experiences/UserWorkExperience';
-import { ExperienceStatus } from '../src/entity/user/experiences/types';
-import {
-  UserJobPreferences,
-  WorkLocationType,
-} from '../src/entity/user/UserJobPreferences';
 import { UserCandidatePreference } from '../src/entity/user/UserCandidatePreference';
 
 jest.mock('../src/common/geo', () => ({
@@ -7340,304 +7334,6 @@ describe('mutation updateNotificationSettings', () => {
       },
       'GRAPHQL_VALIDATION_FAILED',
     );
-  });
-});
-
-describe('user job preferences', () => {
-  const QUERY = `
-    query UserJobPreferences {
-      userJobPreferences {
-        openToOpportunities
-        preferredRoles
-        preferredLocationType
-        openToRelocation
-        currentTotalComp {
-          currency
-          amount
-        }
-      }
-    }
-  `;
-
-  beforeEach(async () => {
-    // Set up a user with job preferences
-    await con.getRepository(UserJobPreferences).save({
-      userId: '1',
-      openToOpportunities: true,
-      preferredRoles: ['Software Engineer', 'Frontend Developer'],
-      preferredLocationType: 'remote',
-      openToRelocation: false,
-      currentTotalComp: {
-        currency: 'USD',
-        amount: 10,
-      },
-    });
-
-    await saveFixtures(con, UserJobPreferences, [
-      {
-        userId: '1',
-        openToOpportunities: true,
-        preferredRoles: ['Software Engineer', 'Frontend Developer'],
-        preferredLocationType: WorkLocationType.Remote,
-        openToRelocation: false,
-        currentTotalComp: {
-          currency: 'USD',
-          amount: 10,
-        },
-      },
-      {
-        userId: '2',
-        openToOpportunities: true,
-        preferredRoles: [],
-        preferredLocationType: WorkLocationType.Remote,
-        openToRelocation: false,
-        currentTotalComp: {
-          currency: 'USD',
-          amount: 10,
-        },
-      },
-    ]);
-
-    // Set up a user with a complete profile
-    await con.getRepository(User).update('1', {
-      flags: {
-        country: 'US',
-      },
-    });
-
-    // Add work experience to make sure profile is complete
-    await saveFixtures(con, UserWorkExperience, [
-      {
-        userId: '1',
-        title: 'Software Engineer',
-        status: ExperienceStatus.Published,
-        description: '',
-        startDate: new Date(),
-      },
-      {
-        userId: '1',
-        title: 'Senior Software Engineer',
-        status: ExperienceStatus.Published,
-        description: '',
-        startDate: new Date(),
-      },
-      {
-        userId: '1',
-        title: 'Software Developer',
-        status: ExperienceStatus.Published,
-        description: '',
-        startDate: new Date(),
-      },
-      {
-        userId: '1',
-        title: 'Frontend Developer',
-        status: ExperienceStatus.Published,
-        description: '',
-        startDate: new Date(),
-      },
-      {
-        userId: '1',
-        title: 'Draft Job Title',
-        status: ExperienceStatus.Draft,
-        description: '',
-        startDate: new Date(),
-      },
-    ]);
-  });
-
-  describe('query userJobPreferences', () => {
-    it('should throw error on query as guest', async () => {
-      return testQueryErrorCode(client, { query: QUERY }, 'UNAUTHENTICATED');
-    });
-
-    it('should return user job preferences', async () => {
-      loggedUser = '1';
-      const res = await client.query(QUERY);
-      expect(res.errors).toBeFalsy();
-      expect(res.data.userJobPreferences).toMatchObject({
-        openToOpportunities: true,
-        preferredRoles: ['Software Engineer', 'Frontend Developer'],
-        preferredLocationType: 'remote',
-        openToRelocation: false,
-        currentTotalComp: {
-          currency: 'USD',
-          amount: 10,
-        },
-      });
-    });
-
-    it('should force openToOpportunities to false if profile is not complete', async () => {
-      loggedUser = '2'; // User without job preferences
-      const res = await client.query(QUERY);
-      expect(res.errors).toBeFalsy();
-      expect(res.data.userJobPreferences).toMatchObject({
-        // fetched as false even if is true on preferences since have no experiences
-        openToOpportunities: false,
-        preferredRoles: [],
-        preferredLocationType: WorkLocationType.Remote,
-        openToRelocation: false,
-        currentTotalComp: {},
-      });
-    });
-  });
-
-  describe('mutation updateUserJobPreferences', () => {
-    const MUTATION = `
-    mutation UpdateUserJobPreferences(
-      $openToOpportunities: Boolean!
-      $preferredRoles: [String!]!
-      $preferredLocationType: String
-      $openToRelocation: Boolean!
-      $currentTotalComp: UserTotalCompensationInput!
-    ) {
-      updateUserJobPreferences(
-        openToOpportunities: $openToOpportunities
-        preferredRoles: $preferredRoles
-        preferredLocationType: $preferredLocationType
-        openToRelocation: $openToRelocation
-        currentTotalComp: $currentTotalComp
-      ) {
-        openToOpportunities
-        preferredRoles
-        preferredLocationType
-        openToRelocation
-        currentTotalComp {
-          currency
-          amount
-        }
-      }
-    }
-  `;
-
-    it('should throw error on mutation as guest', async () => {
-      return testMutationErrorCode(
-        client,
-        {
-          mutation: MUTATION,
-          variables: {
-            openToOpportunities: true,
-            preferredRoles: ['Software Engineer'],
-            preferredLocationType: 'remote',
-            openToRelocation: false,
-            currentTotalComp: {},
-          },
-        },
-        'UNAUTHENTICATED',
-      );
-    });
-
-    it('should update user job preferences', async () => {
-      loggedUser = '1';
-      const variables = {
-        openToOpportunities: false,
-        preferredRoles: ['Product Manager', 'Project Manager'],
-        preferredLocationType: WorkLocationType.Remote,
-        openToRelocation: true,
-        currentTotalComp: {
-          currency: 'EUR',
-          amount: 10,
-        },
-      };
-
-      const res = await client.mutate(MUTATION, { variables });
-      expect(res.errors).toBeFalsy();
-      expect(res.data.updateUserJobPreferences).toMatchObject({
-        openToOpportunities: false,
-        preferredRoles: ['Product Manager', 'Project Manager'],
-        preferredLocationType: WorkLocationType.Remote,
-        openToRelocation: true,
-        currentTotalComp: {
-          currency: 'EUR',
-          amount: 10,
-        },
-      });
-
-      // Verify database state
-      const updatedPrefs = await con
-        .getRepository('UserJobPreferences')
-        .findOne({
-          where: { userId: '1' },
-        });
-      expect(updatedPrefs).toMatchObject({
-        openToOpportunities: false,
-        preferredRoles: ['Product Manager', 'Project Manager'],
-        preferredLocationType: WorkLocationType.Remote,
-        openToRelocation: true,
-        currentTotalComp: {
-          currency: 'EUR',
-          amount: 10,
-        },
-      });
-    });
-
-    it('should throw error on invalid job preferences', async () => {
-      loggedUser = '1';
-
-      // Test with too many preferred roles
-      const variables = {
-        openToOpportunities: true,
-        preferredRoles: [
-          'Role 1',
-          'Role 2',
-          'Role 3',
-          'Role 4',
-          'Role 5',
-          'Role 6',
-        ],
-        preferredLocationType: 'remote',
-        openToRelocation: false,
-        currentTotalComp: {},
-      };
-
-      return testMutationErrorCode(
-        client,
-        { mutation: MUTATION, variables },
-        'GRAPHQL_VALIDATION_FAILED',
-        'Invalid job preferences data. Please check your input.',
-      );
-    });
-
-    it('should not update user job preferences of other user', async () => {
-      // Create job preferences for user 2
-      await con.getRepository('UserJobPreferences').save({
-        userId: '2',
-        openToOpportunities: false,
-        preferredRoles: ['Designer'],
-        preferredLocationType: 'on_site',
-        openToRelocation: true,
-        currentTotalComp: {
-          currency: 'GBP',
-          amount: 20,
-        },
-      });
-
-      // Log in as user 1
-      loggedUser = '1';
-
-      // Try to update user 2's preferences by setting userId
-      const variables = {
-        userId: '2', // This should be ignored
-        openToOpportunities: true,
-        preferredRoles: ['Hacker'],
-        preferredLocationType: 'remote',
-        openToRelocation: false,
-        currentTotalComp: {},
-      };
-
-      const res = await client.mutate(MUTATION, { variables });
-      expect(res.errors).toBeFalsy();
-
-      // Verify that user 1's preferences were updated, not user 2's
-      const user1Prefs = await con.getRepository('UserJobPreferences').findOne({
-        where: { userId: '1' },
-      });
-      expect(user1Prefs?.preferredRoles).toContain('Hacker');
-
-      const user2Prefs = await con.getRepository('UserJobPreferences').findOne({
-        where: { userId: '2' },
-      });
-      expect(user2Prefs?.preferredRoles).toEqual(['Designer']);
-    });
   });
 });
