+++ src/entity/user/experiences/types.ts
@@ -41,3 +41,9 @@ export type ProjectLink = {
   type: ProjectLinkType;
   url: string;
 };
+
+export enum WorkLocationType {
+  Remote = 'remote',
+  Hybrid = 'hybrid',
+  OnSite = 'on_site',
+}
