+++ src/graphorm/index.ts
@@ -36,7 +36,6 @@ import {
   domainOnly,
   getSmartTitle,
   getTranslationRecord,
-  isProfileCompleteById,
   transformDate,
 } from '../common';
 import { GQLComment } from '../schema/comments';
@@ -60,10 +59,6 @@ import {
   ContentPreferenceOrganization,
   ContentPreferenceOrganizationStatus,
 } from '../entity/contentPreference/ContentPreferenceOrganization';
-import {
-  UserCompensation,
-  WorkLocationType,
-} from '../entity/user/UserJobPreferences';
 import { OpportunityUserRecruiter } from '../entity/opportunities/user';
 import { OpportunityUserType } from '../entity/opportunities/types';
 import { OrganizationLinkType } from '../common/schema/organizations';
@@ -1351,49 +1346,6 @@ const obj = new GraphORM({
       },
       seatType: {
         alias: { field: 'status', type: 'string' },
-      },
-    },
-  },
-  UserJobPreferences: {
-    requiredColumns: [
-      'userId',
-      'openToOpportunities',
-      'preferredRoles',
-      'preferredLocationType',
-      'openToRelocation',
-      'currentTotalComp',
-    ],
-    additionalQuery: (ctx, alias, qb) =>
-      qb.andWhere(`"${alias}"."userId" = :currentUserId`, {
-        currentUserId: ctx.userId,
-      }),
-    fields: {
-      openToOpportunities: {
-        transform: async (value: boolean, ctx: Context) => {
-          if (!ctx.userId) {
-            return false;
-          }
-
-          const isProfileComplete = await isProfileCompleteById(
-            ctx.con,
-            ctx.userId,
-          );
-
-          return value && isProfileComplete;
-        },
-      },
-      preferredRoles: {
-        transform: (value: string[] | null) => value || [],
-      },
-      preferredLocationType: {
-        transform: (value: WorkLocationType | null) => value,
-      },
-      openToRelocation: {
-        transform: (value: boolean | null) => value ?? false,
-      },
-      currentTotalComp: {
-        jsonType: true,
-        transform: (value: Partial<UserCompensation> | null) => value || {},
       },
     },
   },
